refactor: update navigation paths and restructure app layout; renamed pantry.jsx component to index.jsx
